Add waks-on-waks-off shuffle

+++ osort/external_memory/algorithm/randomized_shell_sort.hpp
@@ -7,7 +7,7 @@ namespace EM::Algorithm {
 
 class ShellSort {
  public:
-  static const int C = 4;
+  static const int64_t C = 4;
 
   template <typename Iterator>
   static void compareExchange(Iterator i, Iterator j) {
@@ -16,14 +16,14 @@ class ShellSort {
   }
 
   template <typename Iterator>
-  static void compareRegions(Iterator s, Iterator t, int offset) {
-    std::vector<int> mate(offset);
-    for (int count = 0; count < C; count++) {
-      for (int i = 0; i < offset; i++) {
+  static void compareRegions(Iterator s, Iterator t, int64_t offset) {
+    std::vector<int64_t> mate(offset);
+    for (int64_t count = 0; count < C; count++) {
+      for (int64_t i = 0; i < offset; i++) {
         mate[i] = i;
       }
       fisherYatesShuffle(mate.begin(), mate.end());
-      for (int i = 0; i < offset; i++) {
+      for (int64_t i = 0; i < offset; i++) {
         compareExchange(s + i, t + mate[i]);
       }
     }
@@ -31,24 +31,24 @@ class ShellSort {
 
   template <typename Iterator>
   static void randomizedShellSort(Iterator begin, Iterator end) {
-    int n = end - begin;
-    for (int offset = n / 2; offset > 0; offset /= 2) {
-      for (int i = 0; i < n - offset; i += offset) {
+    int64_t n = end - begin;
+    for (int64_t offset = n / 2; offset > 0; offset /= 2) {
+      for (int64_t i = 0; i < n - offset; i += offset) {
         compareRegions(begin + i, begin + i + offset, offset);
       }
-      for (int i = n - offset; i >= offset; i -= offset) {
+      for (int64_t i = n - offset; i >= offset; i -= offset) {
         compareRegions(begin + i - offset, begin + i, offset);
       }
-      for (int i = 0; i < n - 3 * offset; i += offset) {
+      for (int64_t i = 0; i < n - 3 * offset; i += offset) {
         compareRegions(begin + i, begin + i + 3 * offset, offset);
       }
-      for (int i = 0; i < n - 2 * offset; i += offset) {
+      for (int64_t i = 0; i < n - 2 * offset; i += offset) {
         compareRegions(begin + i, begin + i + 2 * offset, offset);
       }
-      for (int i = 0; i < n; i += 2 * offset) {
+      for (int64_t i = 0; i < n; i += 2 * offset) {
         compareRegions(begin + i, begin + i + offset, offset);
       }
-      for (int i = offset; i < n - offset; i += 2 * offset) {
+      for (int64_t i = offset; i < n - offset; i += 2 * offset) {
         compareRegions(begin + i, begin + i + offset, offset);
       }
     }
